Added toggle for AI and player control and added melee ally
Health and hurt/hit box components also added. The melee ally scene is not complete but the the controller is functional and should only require minimal tweaking to get working with the behaviour tree. The state machine from the player should be copy pasted and the states may need to be modified. It will need to be integrated into the behaviour tree.